+++ modules/redos-core/shared/src/main/scala/codes/quine/labo/redos/data/package.scala
@@ -1,0 +1,11 @@
+package codes.quine.labo.redos
+
+/** Package data provides common data structures of this library. */
+package object data {
+
+  /** An alias to unicode.UChar. */
+  type UChar = unicode.UChar
+
+  /** An alias to unicode.UChar. */
+  val UChar = unicode.UChar
+}