+++ modules/redos-core/shared/src/main/scala/codes/quine/labo/redos/data/unicode/IntervalSet.scala
@@ -1,4 +1,4 @@
-package codes.quine.labo.redos.data
+package codes.quine.labo.redos.data.unicode
 
 import scala.math.Ordering.Implicits._