+++ project/CaseMapDataGen.scala
@@ -50,12 +50,9 @@ object CaseMapDataGen extends UnicodeDataGen {
   def source: String = {
     val sb = new mutable.StringBuilder
 
-    sb.append("package codes.quine.labo.redos\n")
-    sb.append("package unicode\n")
+    sb.append("package codes.quine.labo.redos.data.unicode\n")
     sb.append("\n")
     sb.append("import CaseMap.Conversion\n")
-    sb.append("import data.IntervalSet\n")
-    sb.append("import data.UChar\n")
     sb.append("\n")
     sb.append("private[unicode] object CaseMapData {\n")